// frontends/codegen/src/types/project.ts
export interface DbModel {
  id: string
  name: string
  fields: DbField[]
}

export interface DbField {
  id: string
  name: string
  type: string
  isRequired: boolean
  isUnique: boolean
  isArray: boolean
  defaultValue?: string
  relation?: string
}